moved features.js
