+++ documentation/src/pages/index.js
@@ -5,7 +5,7 @@ import Link from '@docusaurus/Link';
 import useDocusaurusContext from '@docusaurus/useDocusaurusContext';
 import useBaseUrl from '@docusaurus/useBaseUrl';
 
-import { features } from './features';
+import { features } from '../features';
 import styles from './styles.module.css';
